moved collection conversion methods into collection.js. Wrote tests for collections.js
