+++ search/lib/convert/collections.js
@@ -1,10 +1,8 @@
-const _ = require('lodash');
 const cmr = require('../cmr/cmr');
-const { wfs, generateAppUrl, extractParam, generateSelfUrl } = require('../util');
+const { wfs, generateAppUrl } = require('../util');
 const { WHOLE_WORLD_BBOX, pointStringToPoints, parseOrdinateString, addPointsToBbox, mergeBoxes } = require('./bounding-box');
-const { cmrGranToFeatureGeoJSON } = require('../cmr/cmr_converter');
 
-const cmrCollSpatialToExtents = (cmrColl) => {
+function cmrCollSpatialToExtents (cmrColl) {
   let bbox = null;
   if (cmrColl.polygons) {
     // an array of arrays of strings containing rings
@@ -17,7 +15,7 @@ const cmrCollSpatialToExtents = (cmrColl) => {
   }
   if (cmrColl.points) {
     // an array of strings of "lat lon"
-    // [ "53.63 -106.2"
+    // [ "53.63 -106.2", "33 -145"]
     const points = cmrColl.points.map(parseOrdinateString);
     bbox = addPointsToBbox(bbox, points);
   }
@@ -25,20 +23,20 @@ const cmrCollSpatialToExtents = (cmrColl) => {
     throw new Error(`Unepxected spatial extent of lines in ${cmrColl.id}`);
   }
   if (cmrColl.boxes) {
-    // a list of strings of south west north east
+    // a list of strings of west north east south
     // TODO can cross antimeridian
     // [ "-14.3601 -176.6525 64.823 151.78372" ]
 
     bbox = cmrColl.boxes.reduce((box, boxStr) => mergeBoxes(box, parseOrdinateString(boxStr)), bbox);
   }
-  if (_.isNull(bbox)) {
+  if (bbox === null) {
     // whole world bbox
     bbox = WHOLE_WORLD_BBOX;
   }
   return bbox;
 };
 
-const stacSearchWithCurrentParams = (event, collId) => {
+function stacSearchWithCurrentParams (event, collId) {
   const newParams = [...event.queryStringParameters] || {};
   newParams.collectionId = collId;
   // The provider param isn't needed once the colleciton id is set.
@@ -46,7 +44,7 @@ const stacSearchWithCurrentParams = (event, collId) => {
   return generateAppUrl(event, '/search/stac', newParams);
 };
 
-const cmrGranuleSearchWithCurrentParams = (event, collId) => {
+function cmrGranuleSearchWithCurrentParams (event, collId) {
   const newParams = [...event.queryStringParameters] || {};
   newParams.collection_concept_id = collId;
   // The provider param isn't needed once the colleciton id is set.
@@ -86,28 +84,9 @@ const cmrCollToWFSColl = (event, cmrColl) => ({
   }
 });
 
-// Move to collection_converter?
-const cmrGranulesToFeatureCollection = (event, cmrGrans) => {
-  const currPage = parseInt(extractParam(event.queryStringParameters, 'page_num', '1'), 10);
-  const nextPage = currPage + 1;
-  const newParams = { ...event.queryStringParameters } || {};
-  newParams.page_num = nextPage;
-  const nextResultsLink = generateAppUrl(event, event.path, newParams);
-
-  return {
-    type: 'FeatureCollection',
-    features: cmrGrans.map(g => cmrGranToFeatureGeoJSON(event, g)),
-    links: {
-      self: generateSelfUrl(event),
-      next: nextResultsLink
-    }
-  };
-};
-
 module.exports = {
   cmrCollSpatialToExtents,
   stacSearchWithCurrentParams,
   cmrGranuleSearchWithCurrentParams,
-  cmrCollToWFSColl,
-  cmrGranulesToFeatureCollection
+  cmrCollToWFSColl
 };
