+++ search/tests/convert/collections.spec.js
@@ -1,60 +1,43 @@
-const { cmrGranulesToFeatureCollection } = require('../../lib/convert');
+const { 
+  cmrCollSpatialToExtents,
+  stacSearchWithCurrentParams,
+  cmrGranulesSearchWithCurrentParams,
+  cmrCollToWFSColl
+} = require('../../lib/convert/collections');
 
-describe('cmrGranulesToFeatureCollection', () => {
-  const cmrGran = [{
-    id: 1,
-    collection_concept_id: 10,
-    dataset_id: 'datasetId',
-    summary: 'summary',
-    time_start: 0,
-    time_end: 1,
-    links: [
-      {
-        href: 'http://example.com/collections/id',
-        rel: 'self',
-        title: 'Info about this collection',
-        type: 'application/json'
+describe('collections', () => {
+  describe('cmrCollSpatialToExtents', () => {
+    // four scenarios need to be tested. The cmrColl object having either polygons, 
+    // points, lines, boxes, or nothing.
+    let cmrCollection;
+
+    it('should return a bounding box from given polygon', () => {
+      cmrCollection = {
+        polygons: [['30 -10 70 33 -45 66']]
+      };
+      expect(cmrCollSpatialToExtents(cmrCollection)).toEqual([-10, 70, 66, -45]);
+    });
+
+    it('should return a bounding box from given points', () => {
+      cmrCollection = {
+        points: ['30 -10', '70 33', '-45 66']
+      };
+      expect(cmrCollSpatialToExtents(cmrCollection)).toEqual([-10, 70, 66, -45]);
+    });
+
+    it('should throw an error if given lines', () => {
+      cmrCollection = {
+        id: 'sampleCollection',
+        lines: [28.2, 44, -44, 109.3]
       }
-    ],
-    data_center: 'USA',
-    points: ['39,77']
-  }];
+      expect(() => { cmrCollSpatialToExtents(cmrCollection); }).toThrow(Error);
+    });
 
-  const event = { headers: { Host: 'example.com' }, queryStringParameters: [] };
-
-  it('should return a cmrGranule to a FeatureCollection', () => {
-    expect(cmrGranulesToFeatureCollection(event, cmrGran)).toEqual({
-      type: 'FeatureCollection',
-      features: [{
-        id: 1,
-        geometry: { type: 'Point', coordinates: [77, 39] },
-        properties: {
-          provider: 'USA',
-          datetime: '0/1'
-        },
-        type: 'Feature',
-        assets: {},
-        links: {
-          self: {
-            rel: 'self',
-            href: 'http://example.com/collections/10/items/1'
-          },
-          parent: {
-            rel: 'parent',
-            href: 'http://example.com/collections/10'
-          },
-          metadata: {
-            href: 'https://cmr.earthdata.nasa.gov/search/concepts/1.native',
-            rel: 'metadata',
-            type: 'application/json',
-            title: undefined
-          }
-        }
-      }],
-      links: {
-        self: 'http://example.com/undefined?',
-        next: 'http://example.com/undefined?page_num=2'
-      }
+    it('should return a bounding box from provided coordinates [west north east south]', () => {
+      cmrCollection = {
+        boxes: ['-23.4 -74.6 54.9 33.3']
+      };
+      expect(cmrCollSpatialToExtents(cmrCollection)).toEqual([-23.4, -74.6, 54.9, 33.3]);
     });
   });
 });
